Add IdleConnTimeout for Transport
Request may be failed if IdleConnTimeout is greater than
IdleConnTimeout of nginx before FIN received.

Close the idle connection before shutdown by nginx server.
